move collators
