+++ mteb/models/model_implementations/clap_models.py
@@ -8,9 +8,9 @@ from tqdm.auto import tqdm
 from transformers import ClapModel, ClapProcessor
 from transformers.modeling_outputs import BaseModelOutputWithPooling
 
-from mteb._create_dataloaders import AudioCollator
 from mteb.models import ModelMeta
 from mteb.models.abs_encoder import AbsEncoder
+from mteb.models.modality_collators import AudioCollator
 
 if TYPE_CHECKING:
     from torch.utils.data import DataLoader
