+++ mteb/models/model_implementations/pe_av_models.py
@@ -6,9 +6,9 @@ import numpy as np
 import torch
 from tqdm.auto import tqdm
 
-from mteb._create_dataloaders import VideoCollator
 from mteb.models import ModelMeta
 from mteb.models.abs_encoder import AbsEncoder
+from mteb.models.modality_collators import VideoCollator
 
 if TYPE_CHECKING:
     from torch.utils.data import DataLoader
