+++ mteb/models/model_implementations/seamlessm4t_models.py
@@ -8,9 +8,9 @@ from tqdm.auto import tqdm
 from transformers import AutoProcessor, SeamlessM4Tv2Model
 
 from mteb import TaskMetadata
-from mteb._create_dataloaders import AudioCollator
 from mteb.models import ModelMeta
 from mteb.models.abs_encoder import AbsEncoder
+from mteb.models.modality_collators import AudioCollator
 from mteb.types import Array, BatchedInput, PromptType
 from mteb.types._encoder_io import AudioInput
